refactor(arch): extract big run_interactive event handlers into ui/run_handlers/
`run_interactive`'s tokio::select! had ~470-line `Done`, ~300-line
`ToolResult`, ~250-line `ContextOverflow`, ~110-line `Interjected`
arms inlined. Each is now a dedicated handler module in
`src/ui/run_handlers/`. The dispatcher arms in run_interactive
collapse to a single function call against a `RunCtx<'a>` bundle.

New module: `src/ui/run_handlers/`
- mod.rs (RunCtx + macro_rules! make_run_ctx! shorthand, 95 lines)
- tool_result.rs (handle_tool_result, 312 lines) — pairs to call,
  renders inside chamber / as `↳` trailer / as fresh chamber for
  parallel-execution races, with colorized edit-diff path
- done.rs (handle_done, 598 lines) — closes chamber, runs plugin
  on-response / on-complete / prepare-next-run (with optional
  agent rebuild on model swap), finalizes streamed text, auto-
  compacts, dispatches decide_post_done_action, spawns background
  review + curator, handles git-worktree return, drains
  interjections
- interjected.rs (handle_interjected, 157 lines) — finalizes
  partial response, persists with Interrupted tool-call entries,
  drains queued interjections
- context_overflow.rs (handle_context_overflow, 290 lines) — auto-
  compact + respawn with tool-side-effect safety + interjection-
  queue draining on each failure branch

RunCtx<'a> bundles 18 fields of mutable borrows + 2 non-mut config
refs the inline handlers were juggling: renderer, session,
response/reasoning buffers + start lines, agent_line_started,
last_tool_name / last_tool_call_id, chamber state, tool_calls
buf + counter, last_collapsed, last_user_prompt, cli, cfg.
Per-handler extras (runner state, channels, feature-gated
managers) are explicit params. The `make_run_ctx!` macro
collapses each call site's RunCtx literal from ~20 lines to one.

Arms that stayed inline (under the 100-line threshold or
control-flow-entangled): Reasoning, Token, ToolCall, ToolStarted,
CustomMessage, Error, TurnStart, TurnEnd, ContextCompacted,
UserMessage, RetryNotice.

Final stats:
- src/ui/mod.rs:   5001 → 3476 lines (−1525, total this branch
  5905 → 3476, −41%)
- run_handlers/:   1414 lines across 5 files
- mod_tests.rs:    571 lines (already extracted prior round)

Build clean, 1718 pass / 0 fail / 0 ignored.

Bonus: removed an unused `use compact_str::CompactString` import
in mod.rs that the handler extraction left behind.
