+++ src/ui/mod.rs
@@ -17,6 +17,7 @@ pub(crate) mod picker;
 #[cfg(feature = "plugin")]
 mod plugin_tree;
 mod renderer;
+mod run_handlers;
 mod search_rewind;
 mod selection;
 mod shell_exec;
@@ -36,7 +37,6 @@ mod tree;
 mod tui;
 mod wrap;
 
-use compact_str::CompactString;
 use crossterm::event;
 use crossterm::event::{KeyCode, KeyModifiers, MouseButton, MouseEventKind};
 use crossterm::style::Color;
@@ -392,6 +392,37 @@ pub async fn run_interactive(
             .to_string(),
     };
     renderer.set_left_panel_info(initial_left_info);
+
+    // Convenience builder for the bundled `RunCtx` borrowed by the
+    // extracted agent-event handlers (`run_handlers::*`). Captures
+    // the live `&mut` refs into the surrounding fn's locals each
+    // time it's expanded. Keeping this as a macro rather than a
+    // helper closure side-steps the multi-borrow lifetime issue —
+    // the closure approach would need to capture every field
+    // by-mut-ref simultaneously, which the borrow checker would
+    // (correctly) reject.
+    macro_rules! make_run_ctx { () => {
+        run_handlers::RunCtx {
+            renderer: &mut renderer,
+            session,
+            response_buf: &mut response_buf,
+            response_start_line: &mut response_start_line,
+            reasoning_buf: &mut reasoning_buf,
+            reasoning_start_line: &mut reasoning_start_line,
+            agent_line_started: &mut agent_line_started,
+            last_tool_name: &mut last_tool_name,
+            last_tool_call_id: &mut last_tool_call_id,
+            tool_chamber_open: &mut tool_chamber_open,
+            chamber_top_start: &mut chamber_top_start,
+            chamber_top_end: &mut chamber_top_end,
+            tool_calls_buf: &mut tool_calls_buf,
+            tool_calls_this_run: &mut tool_calls_this_run,
+            last_collapsed: &mut _last_collapsed,
+            last_user_prompt: &mut last_user_prompt,
+            cli,
+            cfg,
+        }
+    }; }
 
     render_session(&mut renderer, session, cli, cfg, context)?;
     renderer.draw_bottom(
@@ -1739,797 +1770,57 @@ pub async fn run_interactive(
                         // adding a new event variant.
                     }
                     AgentEvent::ToolResult { id, output, .. } => {
-                        // Phase 3: pair the result with its call.
-                        // Prefer id-match; fall back to the most-
-                        // recent Interrupted (pending) entry for
-                        // providers that don't emit ids.
-                        let target = if !id.is_empty() {
-                            tool_calls_buf.iter_mut().rev().find(|e| e.id == id.as_str())
-                        } else {
-                            tool_calls_buf
-                                .iter_mut()
-                                .rev()
-                                .find(|e| matches!(e.state, crate::session::ToolCallState::Interrupted))
-                        };
-                        if let Some(entry) = target {
-                            entry.state = crate::session::ToolCallState::Completed {
-                                result: output.to_string(),
-                            };
-                        }
-                        let show_details = cfg.show_tool_details.unwrap_or(true);
-                        let max_chars = cfg.resolve_tool_result_max_chars();
-                        let show_diff = cfg.resolve_show_edit_diff();
-
-                        // dirge-jzj: if the chamber on screen belongs to a
-                        // DIFFERENT tool call (parallel-execution race
-                        // where ToolResults arrive out of order, or a
-                        // newer ToolCall's TOP displaced this result's
-                        // chamber before the result arrived), paint a
-                        // fresh complete chamber for THIS id below the
-                        // current scroll position. Lets each result land
-                        // in its own correctly-labeled frame regardless
-                        // of completion order. The id-matches case (the
-                        // common sequential path) falls through to the
-                        // existing render paths below.
-                        if !id.is_empty()
-                            && last_tool_call_id.as_deref() != Some(id.as_str())
-                            && show_details
-                        {
-                            // Close whatever chamber is on screen first,
-                            // then paint a fresh TOP for this id. We
-                            // don't reuse the ToolCall handler's TOP-
-                            // paint code path because that fires from a
-                            // different event; the body of the new
-                            // chamber will land via path (a) below now
-                            // that tool_chamber_open=true.
-                            if tool_chamber_open {
-                                close_tool_chamber_passive(
-                                    &mut renderer,
-                                    &mut last_tool_name,
-                                    &mut tool_chamber_open,
-                                    &mut chamber_top_start,
-                                    &mut chamber_top_end,
-                                )?;
-                            }
-                            let (resolved_name, resolved_args) = tool_calls_buf
-                                .iter()
-                                .rev()
-                                .find(|e| e.id == id.as_str())
-                                .map(|e| (e.name.to_string(), e.args.clone()))
-                                .unwrap_or_else(|| (String::new(), serde_json::Value::Null));
-                            if !resolved_name.is_empty() {
-                                let upper = resolved_name.to_ascii_uppercase();
-                                let raw_value =
-                                    format_tool_banner_value(&resolved_name, &resolved_args);
-                                let raw_value = sanitize_output(&raw_value).into_string();
-                                let (frame_w, _) = chamber_widths(&renderer);
-                                let header =
-                                    fit_banner_header(&upper, &raw_value, frame_w);
-                                renderer.write_line("", Color::White)?;
-                                renderer.write_line(&header, c_tool())?;
-                                tool_chamber_open = true;
-                                last_tool_name = Some(resolved_name);
-                                last_tool_call_id = Some(id.to_string());
-                            }
-                            // If the call wasn't in tool_calls_buf (id
-                            // unknown — shouldn't happen post-ToolCall
-                            // but defensive), fall through to path (b)
-                            // trailer; we have no banner to paint.
-                        }
-
-                        // on-tool-end is also fired by HookedToolDyn so the
-                        // host doesn't re-dispatch it here.
-
-                        // Three states at ToolResult time:
-                        //
-                        //  (a) chamber OPEN, show_details=true  → paint body
-                        //      inside the chamber + close with `╰─╯`.
-                        //  (b) chamber NOT OPEN (deny path closed it via
-                        //      the alert handler), show_details=true →
-                        //      emit a single dim `  ↳ {output}` trailer.
-                        //      The trailer is the only thing pinned to
-                        //      the original tool call now that its
-                        //      chamber is gone.
-                        //  (c) show_details=false → no body, but if the
-                        //      chamber is still open we MUST close it
-                        //      (a bare chamber_bottom) or the next
-                        //      output paints inside a dead chamber.
-                        //
-                        // Gating on `tool_chamber_open` (not
-                        // `last_tool_name`) is deliberate: the name
-                        // slot has unrelated clear sites and can drain
-                        // while the chamber TOP is still on screen —
-                        // that's the whole reason for the dedicated
-                        // chamber-state bool.
-                        if !tool_chamber_open && show_details {
-                            // (b) chamber already closed by deny path.
-                            let trimmed = output.trim();
-                            if !trimmed.is_empty() {
-                                let first_line = trimmed.lines().next().unwrap_or("");
-                                renderer.write_line(
-                                    &format!("  ↳ {}", sanitize_output(first_line)),
-                                    theme::dim(),
-                                )?;
-                            }
-                        }
-                        if tool_chamber_open && !show_details {
-                            // (c) chamber on-screen but body suppressed
-                            // — show a single dim "(body hidden)" row
-                            // so the chamber doesn't look like an
-                            // empty box with no content. Then close
-                            // with a bare bottom so a stale `╭─`
-                            // doesn't swallow the next paint.
-                            let (frame_w, inner) = chamber_widths(&renderer);
-                            renderer.write_line(
-                                &chamber_row("(body hidden — show_tool_details=false)", inner),
-                                theme::dim(),
-                            )?;
-                            renderer.write_line(&chamber_bottom(frame_w), theme::dim())?;
-                            tool_chamber_open = false;
-                        }
-                        if tool_chamber_open && show_details {
-                            // Resolve the tool name + banner for the
-                            // collapse store. Prefer the just-stored
-                            // `last_tool_name`; fall back to looking
-                            // up the call by id in `tool_calls_buf`
-                            // (covers paths where `last_tool_name`
-                            // was drained out from under us — same
-                            // shape as the alert-bug fix).
-                            let resolved_name: String = last_tool_name
-                                .clone()
-                                .or_else(|| {
-                                    tool_calls_buf
-                                        .iter()
-                                        .rev()
-                                        .find(|e| e.id == id.as_str())
-                                        .map(|e| e.name.to_string())
-                                })
-                                .unwrap_or_default();
-                            let resolved_args = tool_calls_buf
-                                .iter()
-                                .rev()
-                                .find(|e| e.id == id.as_str())
-                                .map(|e| e.args.clone())
-                                .unwrap_or(serde_json::Value::Null);
-                            let banner_value = format_tool_banner_value(
-                                &resolved_name,
-                                &resolved_args,
-                            );
-                            let max_lines = cfg.resolve_tool_result_max_lines();
-
-                            // Review #7: gate the colorized diff path
-                            // on `resolved_name`, not `last_tool_name`
-                            // — if the name slot drained we'd lose
-                            // the green/red background coloring and
-                            // fall back to plain `render_tool_output`.
-                            let is_edit = resolved_name == "edit" && show_diff;
-                            // Review #6: empty name fallback would
-                            // paint an unnamed chamber AND collapse
-                            // it. Surface a single dim trailer and
-                            // emit the chamber bottom so the chamber
-                            // doesn't orphan. Skip the rest of branch
-                            // (a).
-                            if resolved_name.is_empty() {
-                                let (frame_w, inner) = chamber_widths(&renderer);
-                                let trimmed = output.trim();
-                                let row_text = if trimmed.is_empty() {
-                                    "(unresolved tool, no output)".to_string()
-                                } else {
-                                    let first = trimmed.lines().next().unwrap_or("");
-                                    format!("(unresolved tool) {}", first)
-                                };
-                                renderer.write_line(
-                                    &chamber_row(&row_text, inner),
-                                    theme::dim(),
-                                )?;
-                                renderer.write_line(
-                                    &chamber_bottom(frame_w),
-                                    theme::dim(),
-                                )?;
-                                tool_chamber_open = false;
-                                chamber_top_start = None;
-                                chamber_top_end = None;
-                                last_tool_name = None;
-                                continue;
-                            }
-
-                            if is_edit {
-                                // Colorized diff rendering. The edit tool emits
-                                // its diff block starting with "--- a/<path>" —
-                                // match that exact sentinel to avoid false
-                                // positives on stray "--- " prefixes elsewhere
-                                // in the output.
-                                let lines: Vec<&str> = output.lines().collect();
-                                let diff_start = lines
-                                    .iter()
-                                    .position(|l| l.starts_with("--- a/"));
-                                if let Some(pre) = diff_start {
-                                    let (frame_w, inner) = chamber_widths(&renderer);
-                                    // Pre-diff prose (the edit tool's
-                                    // header line, etc.) renders in
-                                    // the chamber's standard tone.
-                                    for l in &lines[..pre] {
-                                        if !l.is_empty() {
-                                            let txt = sanitize_output(l).into_string();
-                                            renderer.write_line(
-                                                &chamber_row(&txt, inner),
-                                                theme::result(),
-                                            )?;
-                                        }
-                                    }
-                                    // Colorized diff with opencode-style
-                                    // tinted backgrounds: + lines get a
-                                    // dim-green bg (palette 22), - lines
-                                    // get a dim-red bg (palette 52).
-                                    // Header (`--- ` / `+++ ` / `@@`) and
-                                    // context lines have no bg.
-                                    for l in &lines[pre..] {
-                                        let txt = sanitize_output(l).into_string();
-                                        if l.starts_with("--- ") || l.starts_with("+++ ") {
-                                            // Filenames in the diff header get
-                                            // the same accent as section
-                                            // markers elsewhere in chat. Was
-                                            // hardcoded `Color::Cyan` which is
-                                            // invisible on phosphor (same hue
-                                            // as agent text).
-                                            renderer.write_line(
-                                                &chamber_row(&txt, inner),
-                                                theme::accent(),
-                                            )?;
-                                        } else if l.starts_with("@@") {
-                                            // Hunk position markers — use dim
-                                            // so they recede behind the +/-
-                                            // content lines below.
-                                            renderer.write_line(
-                                                &chamber_row(&txt, inner),
-                                                theme::dim(),
-                                            )?;
-                                        } else if l.starts_with('+') {
-                                            renderer.write_line(
-                                                &chamber_row_with_bg(&txt, inner, 22),
-                                                Color::Green,
-                                            )?;
-                                        } else if l.starts_with('-') {
-                                            renderer.write_line(
-                                                &chamber_row_with_bg(&txt, inner, 52),
-                                                Color::Red,
-                                            )?;
-                                        } else {
-                                            renderer.write_line(
-                                                &chamber_row(&txt, inner),
-                                                theme::dim(),
-                                            )?;
-                                        }
-                                    }
-                                    renderer.write_line(
-                                        &chamber_bottom(frame_w),
-                                        theme::dim(),
-                                    )?;
-                                    tool_chamber_open = false;
-                                } else {
-                                    // No diff section found, show normally
-                                    _last_collapsed = render_tool_output(
-                                        &mut renderer,
-                                        &resolved_name,
-                                        &banner_value,
-                                        &output,
-                                        max_chars,
-                                        max_lines,
-                                    )?;
-                                    tool_chamber_open = false;
-                                }
-                            } else {
-                                _last_collapsed = render_tool_output(
-                                    &mut renderer,
-                                    &resolved_name,
-                                    &banner_value,
-                                    &output,
-                                    max_chars,
-                                    max_lines,
-                                )?;
-                                tool_chamber_open = false;
-                            }
-                        }
-                        // Clear after consuming so a future stray ToolResult
-                        // can't be coloured with a stale tool name.
-                        last_tool_name = None;
-                        last_tool_call_id = None;
+                        let mut ctx = make_run_ctx!();
+                        run_handlers::handle_tool_result(
+                            &mut ctx,
+                            id.to_string(),
+                            output.to_string(),
+                        ).await?;
                     }
                     AgentEvent::Done { response, tokens, cost } => {
-                        was_reasoning = false;
-                        // A successful turn must not leave a chamber
-                        // half-painted. If anything slipped through
-                        // — show_details=false skipping the body, an
-                        // in-flight Ask the user resolved with a path
-                        // that didn't reach the bottom paint, etc. —
-                        // close with a plain chamber bottom (not the
-                        // `⚠ tool denied · aborted` wording, which
-                        // would mislead the user about an otherwise-
-                        // successful run).
-                        if tool_chamber_open {
-                            // Same drop-or-close logic as
-                            // close_tool_chamber_passive: if no
-                            // body content was added since the
-                            // TOP was painted (result never
-                            // arrived from the agent — MCP timeout,
-                            // network blip, agent loop bug), drop
-                            // the chamber entirely instead of
-                            // leaving an empty box on screen.
-                            // Otherwise close with a bottom border.
-                            let drop_chamber = match (chamber_top_start, chamber_top_end) {
-                                (Some(_), Some(end)) => renderer.buffer_len() == end,
-                                _ => false,
-                            };
-                            if drop_chamber {
-                                if let Some(start) = chamber_top_start {
-                                    renderer.replace_from(start, Vec::new());
-                                }
-                            } else {
-                                let (frame_w, _) = chamber_widths(&renderer);
-                                renderer.write_line(&chamber_bottom(frame_w), theme::dim())?;
-                            }
-                            tool_chamber_open = false;
-                            chamber_top_start = None;
-                            chamber_top_end = None;
-                        }
-                        last_tool_name = None;
-                        renderer.set_avatar_state(avatar::AvatarState::Done);
-                        #[cfg(feature = "experimental-ui-terminal-tab")]
-                        renderer.set_last_tool_name("");
-
-                        #[allow(unused_mut, unused_variables)]
-                        let mut plugin_followup: Option<String> = None;
-                        #[cfg(feature = "plugin")]
-                        if let Some(pm) = plugin_manager {
-                            let mut mgr = pm.lock().unwrap_or_else(|e| e.into_inner());
-                            match mgr.dispatch(
-                                "on-response",
-                                &format!(
-                                    "@{{:response \"{}\"}}",
-                                    crate::plugin::escape_janet_string(&response)
-                                ),
-                            ) {
-                                Ok(results) if !results.is_empty() => {
-                                    for line in &results {
-                                        // Sanitize plugin output (ANSI injection defense).
-                                        let safe = sanitize_output(line);
-                                        renderer.write_line(
-                                            &format!("[plugin] {}", safe),
-                                            theme::dim(),
-                                        )?;
-                                    }
-                                    plugin_followup = Some(results.join("\n"));
-                                }
-                                Ok(_) => {}
-                                Err(e) => {
-                                    renderer.write_line(
-                                        &format!("[plugin] on-response error: {e}"),
-                                        c_error(),
-                                    )?;
-                                }
-                            }
-                            // Check for pending prompts queued by on-response
-                            if let Some(pending) = mgr.take_pending_prompt() {
-                                plugin_followup = Some(pending);
-                            }
-                            mgr.store_response(&response);
-                            // Fire on-complete after on-response so
-                            // plugins can react to "turn fully done."
-                            // Previously this hook was in HOOK_NAMES
-                            // (so plugins defining it got auto-aliased)
-                            // but no host site dispatched — silent fail.
-                            match mgr.dispatch("on-complete", "@{}") {
-                                Ok(_) => {}
-                                Err(e) => {
-                                    renderer.write_line(
-                                        &format!("[plugin] on-complete error: {e}"),
-                                        c_error(),
-                                    )?;
-                                }
-                            }
-                            // Fire `prepare-next-run` so plugins can
-                            // signal session-level state changes for
-                            // the next run. Closes the gap vs pi's
-                            // `prepareNextTurn` for the auto-apply
-                            // piece: when `harness-next-model` is
-                            // set, the agent is rebuilt with the new
-                            // model RIGHT HERE so the next user
-                            // prompt runs against it without
-                            // requiring `/model X`.
-                            //
-                            // Scope difference vs pi: pi fires
-                            // `prepareNextTurn` between TURNS within
-                            // a single agent run (and can swap model
-                            // mid-stream). dirge fires
-                            // `prepare-next-run` only between RUNS
-                            // (after Done). Mid-stream swap requires
-                            // breaking rig's multi-turn stream and
-                            // restarting with a new agent — that
-                            // would lose partial assistant state, so
-                            // we keep the swap at run boundaries.
-                            match mgr.dispatch("prepare-next-run", "@{}") {
-                                Ok(_) => {}
-                                Err(e) => {
-                                    renderer.write_line(
-                                        &format!("[plugin] prepare-next-run error: {e}"),
-                                        c_error(),
-                                    )?;
-                                }
-                            }
-                            if let Some(next_model) = mgr.take_pending_next_model() {
-                                // Validate: empty string is a
-                                // misconfiguration. Don't replace the
-                                // active model with nothing.
-                                let trimmed = next_model.trim();
-                                if !trimmed.is_empty() && trimmed != session.model.as_str() {
-                                    let new_model_compact = CompactString::new(trimmed);
-                                    let model_obj =
-                                        client.completion_model(new_model_compact.to_string());
-                                    agent = crate::provider::build_agent(
-                                        model_obj,
-                                        cli,
-                                        cfg,
-                                        context,
-                                        permission.clone(),
-                                        ask_tx.clone(),
-                                        question_tx.clone(),
-                                        plan_tx.clone(),
-                                        bg_store.clone(),
-                                        #[cfg(feature = "lsp")]
-                                        lsp_manager.clone(),
-                                        sandbox.clone(),
-                                        #[cfg(feature = "mcp")]
-                                        mcp_manager,
-                                        #[cfg(feature = "semantic")]
-                                        semantic_manager,
-                                    )
-                                    .await;
-                                    let old_model = session.model.clone();
-                                    session.model = new_model_compact.clone();
-                                    session.provider = cli.resolve_provider(cfg);
-                                    // Re-resolve context window for
-                                    // the new model — mirrors the
-                                    // `/model` slash behavior so a
-                                    // 128k→1M jump (or vice versa)
-                                    // updates the status indicator.
-                                    let new_ctx =
-                                        cfg.resolve_context_window(new_model_compact.as_str());
-                                    if new_ctx != session.context_window {
-                                        session.context_window = new_ctx;
-                                    }
-                                    renderer.write_line(
-                                        &format!(
-                                            "[plugin] swapped model: {} → {}",
-                                            old_model, new_model_compact,
-                                        ),
-                                        c_agent(),
-                                    )?;
-                                }
-                            }
-                            // Clear `harness-response` so the next hook
-                            // doesn't see stale text from this turn.
-                            let _ = mgr.eval("(set harness-response nil)");
-                        }
-
-                        if !response_buf.is_empty() {
-                            let max_width = renderer.content_width().saturating_sub(9); // 8-col handle + space
-                            let mut styled = crate::ui::markdown::markdown_to_styled(
-                                &response_buf,
-                                max_width,
-                                c_agent(),
-                            );
-                            if !styled.is_empty() {
-                                styled[0].text =
-                                    CompactString::from(format!("<dirge> {}", styled[0].text));
-                            }
-                            if let Some(start) = response_start_line {
-                                renderer.replace_from(start, styled);
-                                renderer.render_viewport()?;
-                            }
-                        } else if !agent_line_started {
-                            renderer.write("<dirge> ", c_agent())?;
-                        }
-
-                        renderer.write_line("", Color::White)?;
-                        renderer.write_line("", Color::White)?;
-                        // Phase 3: persist structured tool calls
-                        // alongside the assistant text so the next
-                        // resume sees the full tool_use/tool_result
-                        // pairs in convert_history.
-                        session.add_message_with_tool_calls(
-                            MessageRole::Assistant,
-                            &response,
-                            std::mem::take(&mut tool_calls_buf),
-                        );
-                        // TODO(cost-tracking): `tokens` here is the heuristic
-                        // estimate (text.len()/4) and `cost` is always 0.0 —
-                        // these accumulate into placeholder fields and won't
-                        // reflect actual provider usage / billing until we
-                        // pipe rig's `FinalResponse.usage()` through into
-                        // `AgentEvent::Done`. Kept as no-op-ish additions so
-                        // the wiring is in place when real values arrive.
-                        session.total_tokens = session.total_tokens.saturating_add(tokens);
-                        session.total_cost += cost;
-                        // Run ended cleanly — reset the per-run tool-
-                        // call counter so the next user submission
-                        // starts at zero. Mirrored in the Interjected
-                        // branch + both abort paths below.
-                        tool_calls_this_run = 0;
-                        agent_line_started = false;
-                        response_buf.clear();
-                        response_start_line = None;
-                        reasoning_buf.clear();
-                        reasoning_start_line = None;
-
+                        let mut ctx = make_run_ctx!();
                         #[cfg(feature = "loop")]
-                        let loop_running = loop_state.as_ref().is_some_and(|ls| ls.active);
-                        #[cfg(not(feature = "loop"))]
-                        let loop_running = false;
-
-                        if !loop_running
-                            && cfg.resolve_compact_enabled()
-                            && session.needs_compaction(cfg.resolve_reserve_tokens())
-                            && !cli.no_session
-                        {
-                            // Auto-compact failure used to render as a
-                            // single dim red line that scrolled past
-                            // unnoticed — users kept typing into an
-                            // over-full context and saw mysterious
-                            // context-length errors next turn. Frame
-                            // the warning so it visibly stops the eye
-                            // and tells the user what to do next.
-                            renderer.write_line("▒░ auto-compacting context ░▒", theme::accent())?;
-                            let compress_result = handle_compress(
-                                None,
-                                &mut agent, &client, &mut renderer, session, cli, cfg, context,
-                                &permission, &ask_tx, &question_tx, &plan_tx, &bg_store, &sandbox,
-                                #[cfg(feature = "mcp")] mcp_manager,
-                                #[cfg(feature = "semantic")] semantic_manager,
-                                #[cfg(feature = "lsp")] lsp_manager.as_ref(),
-                            ).await;
-                            if let Err(e) = compress_result {
-                                renderer.write_line(
-                                    "╭─ ⚠ AUTO-COMPACT FAILED ─────────────────────────────╮",
-                                    c_error(),
-                                )?;
-                                // Cap the cause length so a sprawling
-                                // multi-line error doesn't blow out the
-                                // box's visual rhythm. The full error
-                                // is still in the agent's recovery
-                                // path; this is for the user-facing
-                                // hint only.
-                                let cause = {
-                                    let s = e.to_string().replace('\n', " ");
-                                    if s.chars().count() > 64 {
-                                        let mut out: String = s.chars().take(63).collect();
-                                        out.push('…');
-                                        out
-                                    } else {
-                                        s
-                                    }
-                                };
-                                renderer.write_line(
-                                    &format!("│ cause: {}", cause),
-                                    c_error(),
-                                )?;
-                                renderer.write_line(
-                                    "│ context is over the threshold — replies may start",
-                                    c_error(),
-                                )?;
-                                renderer.write_line(
-                                    "│ hitting context-length errors. Try /compress",
-                                    c_error(),
-                                )?;
-                                renderer.write_line(
-                                    "│ manually, /clear to start fresh, or restart with",
-                                    c_error(),
-                                )?;
-                                renderer.write_line(
-                                    "│ a larger context_window in config.",
-                                    c_error(),
-                                )?;
-                                renderer.write_line(
-                                    "╰─────────────────────────────────────────────────────╯",
-                                    c_error(),
-                                )?;
-                            }
-                        }
-
-                        if !cli.no_session
-                            && let Err(e) = crate::session::storage::save_session(session)
-                        {
-                            renderer.write_line(
-                                &format!("warning: failed to save session: {}", e),
-                                c_error(),
-                            )?;
-                        }
-                        is_running = false;
-                        if let Some(h) = agent_abort.take() { h.abort(); }
-                        agent_rx = None;
-                        agent_interject = None;
-
-                        #[cfg(feature = "plugin")]
-                        let followup_for_decision = plugin_followup.clone();
-                        #[cfg(not(feature = "plugin"))]
-                        let followup_for_decision: Option<String> = None;
-
-                        #[cfg(feature = "loop")]
-                        let (loop_active, loop_should_stop) = loop_state
-                            .as_ref()
-                            .map(|ls| (ls.active, ls.active && ls.should_stop()))
-                            .unwrap_or((false, false));
-                        #[cfg(not(feature = "loop"))]
-                        let (loop_active, loop_should_stop) = (false, false);
-
-                        let action = crate::plugin::decide_post_done_action(
-                            followup_for_decision,
-                            loop_active,
-                            loop_should_stop,
-                        );
-
-                        match action {
-                            crate::plugin::PostDoneAction::Followup(text) => {
-                                let followup_prompt = text + "\n\nContinue.";
-                                last_user_prompt.clone_from(&followup_prompt);
-                                let runner = agent.clone().spawn_runner(
-                                    crate::agent::tools::background::prepend_pending_notifications(&followup_prompt, bg_store.as_ref()),
-                                    crate::agent::runner::convert_history(session),
-                                    Some(interjection_queue.clone()),
-                                );
-                                agent_rx = Some(runner.event_rx);
-                                                agent_abort = Some(runner.task);
-                                                agent_interject = Some(runner.interject_tx);
-                                is_running = true;
-                            }
-                            crate::plugin::PostDoneAction::LoopStop => {
-                                #[cfg(feature = "loop")]
-                                if let Some(ref mut ls) = loop_state {
-                                    renderer.write_line(
-                                        &format!(
-                                            "[loop] max iterations ({}) reached, stopping",
-                                            ls.iteration
-                                        ),
-                                        c_agent(),
-                                    )?;
-                                    ls.active = false;
-                                    loop_label = None;
-                                }
-                            }
-                            crate::plugin::PostDoneAction::LoopIter => {
-                                #[cfg(feature = "loop")]
-                                if let Some(ref mut ls) = loop_state {
-                                    let summary: String = response.chars().take(200).collect();
-                                    ls.last_summary = Some(summary);
-                                    ls.iteration += 1;
-                                    let prompt = ls.build_prompt();
-                                    last_user_prompt.clone_from(&prompt);
-                                    let runner = agent.clone().spawn_runner(
-                                        crate::agent::tools::background::prepend_pending_notifications(&prompt, bg_store.as_ref()),
-                                        Vec::new(),
-                                        Some(interjection_queue.clone()),
-                                    );
-                                    agent_rx = Some(runner.event_rx);
-                                                agent_abort = Some(runner.task);
-                                                agent_interject = Some(runner.interject_tx);
-                                    is_running = true;
-                                    loop_label = Some(ls.iteration_label());
-                                    renderer.write_line(
-                                        &format!("[loop] launching {}", ls.iteration_label()),
-                                        c_agent(),
-                                    )?;
-                                }
-                            }
-                            crate::plugin::PostDoneAction::Idle => {}
-                        }
-
-                        // Phase 4: spawn background review when the
-                        // session is truly idle (no plugin followup,
-                        // loop iteration, or worktree cleanup claimed
-                        // the next turn). Fire-and-forget — the review
-                        // runs in a tokio task and never blocks the user.
-                        if !is_running {
-                            let cwd = std::env::current_dir().unwrap_or_else(|_| ".".into());
-                            let paths = crate::extras::dirge_paths::ProjectPaths::new(&cwd);
-
-                            // Persist the completed turn to the SQLite
-                            // session DB for future search. Uses a
-                            // stable session id so messages from the
-                            // same interactive session are grouped.
-                            // Includes tool names + results for FTS5.
-                            persist_turn_to_db(session, &last_user_prompt, &response, &tool_calls_buf);
-
-                            let transcript = crate::agent::review::build_transcript(session);
-                            crate::agent::review::spawn_background_review(
-                                agent.clone(),
-                                paths.clone(),
-                                transcript,
-                                None,
-                            );
-
-                            // Curator check: run periodic skill maintenance
-                            // if the interval has elapsed. Fire-and-forget.
-                            tokio::spawn(async move {
-                                if let Ok(mut curator) = crate::extras::skills::curator::Curator::new(&paths) {
-                                    if curator.should_run_now() {
-                                        let _ = tokio::task::spawn_blocking(move || {
-                                            let _ = curator.apply_automatic_transitions();
-                                        }).await;
-                                    }
-                                }
-                            });
-                        }
-
+                        let loop_bits = run_handlers::done::LoopBits {
+                            state: &mut loop_state,
+                            label: &mut loop_label,
+                        };
                         #[cfg(feature = "git-worktree")]
-                        if let Some(main_path) = wt_return_path.take() {
-                            match std::env::set_current_dir(&main_path) {
-                                Ok(()) => {
-                                    session.working_dir = compact_str::CompactString::new(&main_path);
-                                    context.reload();
-                                    let model = client.completion_model(session.model.to_string());
-                                    agent = crate::provider::build_agent(
-                                        model,
-                                        cli,
-                                        cfg,
-                                        context,
-                                        permission.clone(),
-                                        ask_tx.clone(),
-                                        question_tx.clone(),
-                                        plan_tx.clone(),
-                                        bg_store.clone(),
-                                                                                #[cfg(feature = "lsp")]
-                                                                                lsp_manager.clone(),
-                                        sandbox.clone(),
-                                        #[cfg(feature = "mcp")] mcp_manager,
-                                        #[cfg(feature = "semantic")] semantic_manager,
-                                    ).await;
-                                    render_session(&mut renderer, session, cli, cfg, context)?;
-                                    renderer.write_line(
-                                        &format!("merged and returned to main repo at {}", main_path),
-                                        c_agent(),
-                                    )?;
-                                }
-                                Err(e) => {
-                                    renderer.write_line(
-                                        &format!("warning: failed to change back to main repo: {}", e),
-                                        c_error(),
-                                    )?;
-                                }
-                            }
-                        }
-
-                        // Drain the interjection queue once the run is fully
-                        // idle (no plugin follow-up, loop iteration, or worktree
-                        // cleanup grabbed the next turn). Concatenate all
-                        // queued messages as a single new user turn and kick
-                        // off another run against the now-stable agent/cwd.
-                        if !is_running && !interjection_queue.lock().unwrap().is_empty() {
-                            let queued: Vec<String> = interjection_queue.lock().unwrap().drain(..).collect();
-                            let combined = queued.join("\n\n");
-                            // No write_user_lines here — the loop's
-                            // MessageStart{User} → AgentEvent::UserMessage
-                            // bridge will render the user's text once,
-                            // post-stripping the system-reminder block.
-                            // Calling write_user_lines here would
-                            // duplicate the render (see commit 7584bdf
-                            // for the original regular-input fix).
-
-                            last_user_prompt.clone_from(&combined);
-                            let history = crate::agent::runner::convert_history(session);
-                            session.add_message(MessageRole::User, &combined);
-
-                            let runner = agent.clone().spawn_runner(
-                                crate::agent::tools::background::prepend_pending_notifications(
-                                    &combined,
-                                    bg_store.as_ref(),
-                                ),
-                                history,
-                                Some(interjection_queue.clone()),
-                            );
-                            agent_rx = Some(runner.event_rx);
-                            agent_abort = Some(runner.task);
-                            agent_interject = Some(runner.interject_tx);
-                            is_running = true;
-                        }
+                        let worktree_bits = run_handlers::done::WorktreeBits {
+                            return_path: &mut wt_return_path,
+                        };
+                        run_handlers::handle_done(
+                            &mut ctx,
+                            response,
+                            tokens,
+                            cost,
+                            &mut was_reasoning,
+                            &mut is_running,
+                            &mut agent,
+                            &client,
+                            context,
+                            &permission,
+                            &ask_tx,
+                            &question_tx,
+                            &plan_tx,
+                            &bg_store,
+                            &sandbox,
+                            &mut agent_rx,
+                            &mut agent_abort,
+                            &mut agent_interject,
+                            &interjection_queue,
+                            #[cfg(feature = "mcp")]
+                            mcp_manager,
+                            #[cfg(feature = "semantic")]
+                            semantic_manager,
+                            #[cfg(feature = "lsp")]
+                            lsp_manager.as_ref(),
+                            #[cfg(feature = "plugin")]
+                            plugin_manager,
+                            #[cfg(feature = "loop")]
+                            loop_bits,
+                            #[cfg(feature = "git-worktree")]
+                            worktree_bits,
+                        ).await?;
                     }
                     #[cfg(feature = "plugin")]
                     AgentEvent::CustomMessage { payload } => {
@@ -2564,158 +1855,31 @@ pub async fn run_interactive(
                         let _ = payload;
                     }
                     AgentEvent::Interjected { partial_response, tokens } => {
-                        was_reasoning = false;
-                        close_tool_chamber_if_open(&mut renderer, &mut last_tool_name, &mut tool_chamber_open)?;
-
-                        // Finalize whatever assistant text streamed so far so
-                        // the conversation history reflects what the user saw,
-                        // not a phantom turn that "never happened".
-                        if !response_buf.is_empty() {
-                            let max_width = renderer.content_width().saturating_sub(9); // 8-col handle + space
-                            let mut styled = crate::ui::markdown::markdown_to_styled(
-                                &response_buf,
-                                max_width,
-                                c_agent(),
-                            );
-                            if !styled.is_empty() {
-                                styled[0].text =
-                                    CompactString::from(format!("<dirge> {}", styled[0].text));
-                            }
-                            if let Some(start) = response_start_line {
-                                renderer.replace_from(start, styled);
-                                renderer.render_viewport()?;
-                            }
-                        }
-                        renderer.write_line("", Color::White)?;
-                        renderer.write_line(
-                            "(interjected — stopped at last tool-result boundary)",
-                            theme::dim(),
-                        )?;
-                        renderer.write_line("", Color::White)?;
-
-                        // Record the (partial) assistant response in session
-                        // history. Even truncated, it lets the LLM see what
-                        // it had said when the user spoke up.
-                        if !partial_response.is_empty() {
-                            // Persist the partial turn to session DB
-                            // before tool_calls_buf is consumed.
-                            persist_turn_to_db(session, &last_user_prompt, &partial_response, &tool_calls_buf);
-
-                            // Phase 3: same structured persistence
-                            // as the Done branch. Any pending entries
-                            // (tool calls without a result yet) keep
-                            // their Interrupted state — the LLM
-                            // sees [Tool execution was interrupted]
-                            // tool_result on resume.
-                            session.add_message_with_tool_calls(
-                                MessageRole::Assistant,
-                                &partial_response,
-                                std::mem::take(&mut tool_calls_buf),
-                            );
-                            // TODO(cost-tracking): same caveat as the Done
-                            // branch — `tokens` is an estimate, not actual
-                            // provider usage. Wire after rig usage plumbing.
-                            session.total_tokens = session.total_tokens.saturating_add(tokens);
-                        } else {
-                            // No partial text but maybe pending tool
-                            // calls — drop them; the session already
-                            // captured them via prior turns or they
-                            // were a single-call abort with no text.
-                            tool_calls_buf.clear();
-                        }
-                        // Run ended (interjection-style) — reset the
-                        // per-run tool-call counter alongside the
-                        // other per-run state.
-                        tool_calls_this_run = 0;
-                        agent_line_started = false;
-                        response_buf.clear();
-                        response_start_line = None;
-                        reasoning_buf.clear();
-                        reasoning_start_line = None;
-
-                        if !cli.no_session
-                            && let Err(e) = crate::session::storage::save_session(session)
-                        {
-                            renderer.write_line(
-                                &format!("warning: failed to save session: {}", e),
-                                c_error(),
-                            )?;
-                        }
-                        is_running = false;
-                        if let Some(h) = agent_abort.take() { h.abort(); }
-                        agent_rx = None;
-                        agent_interject = None;
-
-                        // Drain the queue immediately — it's guaranteed to be
-                        // non-empty here since the runner only emits this
-                        // event when the UI signaled an interjection, and the
-                        // signal is only sent from the queue-push code path.
-                        if !interjection_queue.lock().unwrap().is_empty() {
-                            let queued: Vec<String> = interjection_queue.lock().unwrap().drain(..).collect();
-                            let combined = queued.join("\n\n");
-                            // No write_user_lines — same reasoning as
-                            // the idle-drain path above; the loop's
-                            // UserMessage bridge handles the render.
-
-                            last_user_prompt.clone_from(&combined);
-                            let history = crate::agent::runner::convert_history(session);
-                            session.add_message(MessageRole::User, &combined);
-
-                            let runner = agent.clone().spawn_runner(
-                                crate::agent::tools::background::prepend_pending_notifications(
-                                    &combined,
-                                    bg_store.as_ref(),
-                                ),
-                                history,
-                                Some(interjection_queue.clone()),
-                            );
-                            agent_rx = Some(runner.event_rx);
-                            agent_abort = Some(runner.task);
-                            agent_interject = Some(runner.interject_tx);
-                            is_running = true;
-                        }
+                        let mut ctx = make_run_ctx!();
+                        run_handlers::handle_interjected(
+                            &mut ctx,
+                            partial_response,
+                            tokens,
+                            &mut was_reasoning,
+                            &mut is_running,
+                            &agent,
+                            &mut agent_rx,
+                            &mut agent_abort,
+                            &mut agent_interject,
+                            &interjection_queue,
+                            &bg_store,
+                        ).await?;
                     }
                     AgentEvent::ContextOverflow { prompt, error } => {
-                        // Audit H17: the streaming run hit a context-
-                        // length error. Auto-compact then re-spawn with
-                        // the same prompt against the now-compacted
-                        // history — opencode-style automatic recovery
-                        // (compaction.ts:477-558) instead of leaving the
-                        // user stranded at the error.
-                        was_reasoning = false;
-                        close_tool_chamber_if_open(&mut renderer, &mut last_tool_name, &mut tool_chamber_open)?;
-                        let safe = sanitize_output(&error);
-                        renderer.write_line(
-                            &format!("context overflow: {}", safe),
-                            c_error(),
-                        )?;
-                        // Persist what we have so far (partial response
-                        // + tool calls) before tearing down the runner.
-                        persist_turn_to_db(session, &last_user_prompt, &response_buf, &tool_calls_buf);
-                        // Tear down the current runner before respawn.
-                        if let Some(h) = agent_abort.take() {
-                            h.abort();
-                        }
-                        agent_rx = None;
-                        agent_interject = None;
-                        agent_line_started = false;
-                        response_buf.clear();
-                        response_start_line = None;
-                        reasoning_buf.clear();
-                        reasoning_start_line = None;
-
-                        renderer.write_line(
-                            "▒░ auto-compacting then retrying ░▒",
-                            theme::accent(),
-                        )?;
-                        let compress_result = handle_compress(
-                            None,
+                        let mut ctx = make_run_ctx!();
+                        run_handlers::handle_context_overflow(
+                            &mut ctx,
+                            prompt,
+                            error,
+                            &mut was_reasoning,
+                            &mut is_running,
                             &mut agent,
                             &client,
-                            &mut renderer,
-                            session,
-                            cli,
-                            cfg,
                             context,
                             &permission,
                             &ask_tx,
@@ -2723,143 +1887,17 @@ pub async fn run_interactive(
                             &plan_tx,
                             &bg_store,
                             &sandbox,
+                            &mut agent_rx,
+                            &mut agent_abort,
+                            &mut agent_interject,
+                            &interjection_queue,
                             #[cfg(feature = "mcp")]
                             mcp_manager,
                             #[cfg(feature = "semantic")]
                             semantic_manager,
                             #[cfg(feature = "lsp")]
                             lsp_manager.as_ref(),
-                        )
-                        .await;
-
-                        // Review #1: compress can return Ok WITHOUT
-                        // shrinking the session (three no-op paths
-                        // inside `handle_compress`). Respawning
-                        // against the unchanged history just re-emits
-                        // ContextOverflow and infinite-loops the
-                        // auto-recovery. Only respawn on `Compacted`.
-                        //
-                        // Review #2: re-issuing the same prompt
-                        // re-runs any side-effecting tool calls the
-                        // failed turn already made. The interactive
-                        // retry loop already refuses to retry when
-                        // `had_tool_calls`; the auto path used to
-                        // bypass that safety. We have no direct
-                        // `had_tool_calls` signal here (the runner
-                        // emitted ContextOverflow without telling us
-                        // whether tools fired). Approximate it by
-                        // comparing `tool_calls_this_run > 0`, which
-                        // tracks every ToolCall event observed during
-                        // the failed turn. If any tool ran, surface
-                        // the error and let the user decide.
-                        let tools_already_ran = tool_calls_this_run > 0;
-                        // Reset the abort-trailer counter regardless
-                        // — the failed run is over.
-                        tool_calls_this_run = 0;
-                        match compress_result {
-                            Ok(crate::ui::slash::CompressOutcome::Compacted)
-                                if !tools_already_ran =>
-                            {
-                                // Build history from the compacted session.
-                                // Drop the trailing User message because
-                                // it's the prompt we're about to resubmit
-                                // — otherwise rig would receive it twice.
-                                let mut history = crate::agent::runner::convert_history(session);
-                                if let Some(last) = history.last()
-                                    && matches!(last, rig::completion::Message::User { .. })
-                                {
-                                    history.pop();
-                                }
-                                let prompt_owned = prompt.to_string();
-                                last_user_prompt.clone_from(&prompt_owned);
-                                let prepared_prompt =
-                                    crate::agent::tools::background::prepend_pending_notifications(
-                                        &prompt_owned,
-                                        bg_store.as_ref(),
-                                    );
-                                let runner =
-                                    agent.clone().spawn_runner(prepared_prompt, history, Some(interjection_queue.clone()));
-                                agent_rx = Some(runner.event_rx);
-                                agent_abort = Some(runner.task);
-                                agent_interject = Some(runner.interject_tx);
-                                is_running = true;
-                                // Review #4: collapsed result from the
-                                // failed run is stale — the user will
-                                // care about results from the new
-                                // attempt, not what got truncated
-                                // before the overflow.
-                                _last_collapsed = None;
-                                renderer.write_line(
-                                    "  ↳ resumed run with compacted history",
-                                    theme::dim(),
-                                )?;
-                            }
-                            Ok(crate::ui::slash::CompressOutcome::Compacted) => {
-                                // Compacted, but tool side-effects
-                                // already applied — refusing auto-
-                                // retry. User can re-issue manually.
-                                renderer.write_line(
-                                    "  ↳ context compacted, but the failed run already invoked tools — not auto-retrying. Re-issue your prompt manually if you want to continue.",
-                                    c_error(),
-                                )?;
-                                is_running = false;
-                                let dropped = interjection_queue.lock().unwrap().len();
-                                interjection_queue.lock().unwrap().clear();
-                                if dropped > 0 {
-                                    renderer.write_line(
-                                        &format!(
-                                            "{} queued message{} dropped due to tool-side-effect safety",
-                                            dropped,
-                                            if dropped == 1 { "" } else { "s" }
-                                        ),
-                                        c_error(),
-                                    )?;
-                                }
-                            }
-                            Ok(crate::ui::slash::CompressOutcome::NoOp { reason }) => {
-                                renderer.write_line(
-                                    &format!(
-                                        "auto-compact made no progress ({reason}); leaving session as-is. Try /compress with stricter instructions, lower keep_recent_tokens, or /clear."
-                                    ),
-                                    c_error(),
-                                )?;
-                                is_running = false;
-                                let dropped = interjection_queue.lock().unwrap().len();
-                                interjection_queue.lock().unwrap().clear();
-                                if dropped > 0 {
-                                    renderer.write_line(
-                                        &format!(
-                                            "{} queued message{} dropped due to compact no-op",
-                                            dropped,
-                                            if dropped == 1 { "" } else { "s" }
-                                        ),
-                                        c_error(),
-                                    )?;
-                                }
-                            }
-                            Err(ce) => {
-                                renderer.write_line(
-                                    &format!(
-                                        "auto-compact failed ({}); leaving session as-is. Try /compress manually or /clear.",
-                                        ce
-                                    ),
-                                    c_error(),
-                                )?;
-                                is_running = false;
-                                let dropped = interjection_queue.lock().unwrap().len();
-                                interjection_queue.lock().unwrap().clear();
-                                if dropped > 0 {
-                                    renderer.write_line(
-                                        &format!(
-                                            "{} queued message{} dropped due to compact failure",
-                                            dropped,
-                                            if dropped == 1 { "" } else { "s" }
-                                        ),
-                                        c_error(),
-                                    )?;
-                                }
-                            }
-                        }
+                        ).await?;
                     }
                     AgentEvent::Error(e) => {
                         was_reasoning = false;
@@ -4434,568 +3472,5 @@ pub async fn run_interactive(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::ui::search_rewind::update_search;
-    use unicode_width::UnicodeWidthStr;
-
-    // ============================================================
-    // apply_subagent_panel_event — left-panel cleanup
-    // ============================================================
-
-    use crate::agent::tools::task::SubagentChatEvent as E;
-
-    /// Spawn → row appears in "running" state with the prompt.
-    #[test]
-    fn subagent_panel_spawn_inserts_running_row() {
-        let mut rows = indexmap::IndexMap::new();
-        apply_subagent_panel_event(
-            &mut rows,
-            &E::Spawn {
-                id: "abc123".into(),
-                prompt: "build the binary".into(),
-            },
-        );
-        assert_eq!(rows.len(), 1);
-        let (state, prompt, _files) = rows.get("abc123").unwrap();
-        assert_eq!(state, "running");
-        assert_eq!(prompt, "build the binary");
-    }
-
-    /// Complete → row is REMOVED (the bug being fixed). Previously
-    /// the row's state changed to "completed" and the entry stayed
-    /// in the map forever, accumulating stale ✓ glyphs in the panel.
-    #[test]
-    fn subagent_panel_complete_removes_row() {
-        let mut rows = indexmap::IndexMap::new();
-        apply_subagent_panel_event(
-            &mut rows,
-            &E::Spawn {
-                id: "abc123".into(),
-                prompt: "build the binary".into(),
-            },
-        );
-        apply_subagent_panel_event(
-            &mut rows,
-            &E::Complete {
-                id: "abc123".into(),
-                result: "ok".into(),
-            },
-        );
-        assert!(rows.is_empty(), "completed subagent must be removed");
-    }
-
-    /// Failed → row is REMOVED (same cleanup contract as Complete).
-    #[test]
-    fn subagent_panel_failed_removes_row() {
-        let mut rows = indexmap::IndexMap::new();
-        apply_subagent_panel_event(
-            &mut rows,
-            &E::Spawn {
-                id: "xyz789".into(),
-                prompt: "run tests".into(),
-            },
-        );
-        apply_subagent_panel_event(
-            &mut rows,
-            &E::Failed {
-                id: "xyz789".into(),
-                error: "boom".into(),
-            },
-        );
-        assert!(rows.is_empty(), "failed subagent must be removed");
-    }
-
-    /// Mixed: several spawns + one completion leaves the rest in
-    /// place and preserves insertion order (oldest at top).
-    #[test]
-    fn subagent_panel_mixed_lifecycle_preserves_order() {
-        let mut rows = indexmap::IndexMap::new();
-        for id in ["a", "b", "c"] {
-            apply_subagent_panel_event(
-                &mut rows,
-                &E::Spawn {
-                    id: id.into(),
-                    prompt: format!("task {id}"),
-                },
-            );
-        }
-        // Remove the middle one.
-        apply_subagent_panel_event(
-            &mut rows,
-            &E::Complete {
-                id: "b".into(),
-                result: "ok".into(),
-            },
-        );
-        assert_eq!(rows.len(), 2);
-        let remaining: Vec<&str> = rows.keys().map(String::as_str).collect();
-        assert_eq!(
-            remaining,
-            vec!["a", "c"],
-            "shift_remove must preserve insertion order of survivors"
-        );
-    }
-
-    /// Complete/Failed for an unknown id is a no-op (defensive —
-    /// shouldn't happen since Complete always follows Spawn, but if
-    /// the event ordering ever drifts, don't panic).
-    #[test]
-    fn subagent_panel_complete_unknown_id_is_noop() {
-        let mut rows = indexmap::IndexMap::new();
-        apply_subagent_panel_event(
-            &mut rows,
-            &E::Complete {
-                id: "never-spawned".into(),
-                result: "ok".into(),
-            },
-        );
-        assert!(rows.is_empty());
-    }
-
-    /// dirge-bfd: Ctrl-F search uses fuzzy matching (nucleo) — typos,
-    /// non-contiguous subsequences, and missing characters all match
-    /// where they wouldn't under the prior substring scheme.
-    #[test]
-    fn fuzzy_search_matches_non_contiguous_subsequence() {
-        let mut renderer = crate::ui::renderer::Renderer::new().expect("renderer");
-        renderer
-            .write_line("connect to database", Color::White)
-            .unwrap();
-        renderer
-            .write_line("contributing guide", Color::White)
-            .unwrap();
-        renderer
-            .write_line("totally unrelated", Color::White)
-            .unwrap();
-
-        let mut matches: Vec<usize> = Vec::new();
-        let mut selected = 0;
-
-        // Substring `ctd` matches nothing under the old `contains`
-        // scheme. Fuzzy matches "connect to database" by its
-        // c-o-n-n-e-C-T-o-D... subsequence.
-        update_search(&renderer, "ctd", &mut matches, &mut selected);
-        assert!(
-            !matches.is_empty(),
-            "fuzzy `ctd` should produce matches; matches={matches:?}",
-        );
-
-        // Empty / whitespace queries clear matches.
-        update_search(&renderer, "", &mut matches, &mut selected);
-        assert!(matches.is_empty());
-        update_search(&renderer, "   ", &mut matches, &mut selected);
-        assert!(matches.is_empty());
-
-        // Lowercase query matches (smart case).
-        update_search(&renderer, "database", &mut matches, &mut selected);
-        assert!(matches.iter().any(|&i| {
-            renderer
-                .buffer_lines()
-                .get(i)
-                .map(|s| s.contains("database"))
-                .unwrap_or(false)
-        }));
-    }
-
-    /// dirge-bfd: Ctrl-F search uses fuzzy matching (nucleo) — typos,
-    /// structured entries on the stashed message. Pending entries
-    /// stay Interrupted (no matching result arrived); on resume,
-    /// `convert_history` will emit a [Tool execution was
-    /// interrupted] tool_result so the LLM sees paired blocks.
-    #[test]
-    fn capture_partial_on_abort_preserves_pending_tool_calls_as_interrupted() {
-        let mut session = crate::session::Session::new("p", "m", 100_000);
-        let mut buf = String::from("Running bash...");
-        let mut calls = vec![
-            crate::session::ToolCallEntry {
-                id: "tc_abc".to_string(),
-                name: "bash".to_string(),
-                args: serde_json::json!({"cmd": "sleep 99"}),
-                state: crate::session::ToolCallState::Interrupted,
-            },
-            crate::session::ToolCallEntry {
-                id: "tc_xyz".to_string(),
-                name: "read".to_string(),
-                args: serde_json::json!({"path": "/etc/hostname"}),
-                state: crate::session::ToolCallState::Completed {
-                    result: "myhost".to_string(),
-                },
-            },
-        ];
-        let stashed = capture_partial_on_abort(&mut buf, &mut session, "Ctrl+C", 2, &mut calls);
-        assert!(stashed);
-        assert!(calls.is_empty(), "tool_calls_buf must be drained on stash");
-
-        let last = session.messages.last().unwrap();
-        assert_eq!(last.tool_calls.len(), 2);
-        let interrupted = last
-            .tool_calls
-            .iter()
-            .find(|e| e.id == "tc_abc")
-            .expect("missing interrupted entry");
-        assert!(matches!(
-            interrupted.state,
-            crate::session::ToolCallState::Interrupted,
-        ));
-        let completed = last
-            .tool_calls
-            .iter()
-            .find(|e| e.id == "tc_xyz")
-            .expect("missing completed entry");
-        match &completed.state {
-            crate::session::ToolCallState::Completed { result } => {
-                assert_eq!(result, "myhost");
-            }
-            other => panic!("expected Completed; got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn capture_partial_on_abort_stashes_partial_with_trailer() {
-        let mut session = crate::session::Session::new("openrouter", "test-model", 100_000);
-        let baseline = session.messages.len();
-        let mut buf = String::from("I was about to explain that");
-        let stashed =
-            capture_partial_on_abort(&mut buf, &mut session, "Ctrl+C", 0, &mut Vec::new());
-        assert!(stashed);
-        assert_eq!(session.messages.len(), baseline + 1);
-        let last = session.messages.last().unwrap();
-        assert_eq!(last.role, crate::session::MessageRole::Assistant);
-        assert!(
-            last.content.contains("I was about to explain that"),
-            "must keep the original partial: {:?}",
-            last.content,
-        );
-        assert!(
-            last.content.contains("[interrupted by user (Ctrl+C)]"),
-            "must include the interruption trailer: {:?}",
-            last.content,
-        );
-        assert!(buf.is_empty(), "buf must be cleared after stash");
-    }
-
-    // Aborting when nothing has streamed yet is a no-op — we don't
-    // want a session full of empty "[interrupted]" messages from
-    // mistaken Ctrl+C presses.
-    #[test]
-    fn capture_partial_on_abort_noop_on_empty_buf() {
-        let mut session = crate::session::Session::new("openrouter", "test-model", 100_000);
-        let baseline = session.messages.len();
-        let mut buf = String::new();
-        let stashed =
-            capture_partial_on_abort(&mut buf, &mut session, "Ctrl+C", 0, &mut Vec::new());
-        assert!(!stashed);
-        assert_eq!(session.messages.len(), baseline);
-    }
-
-    // Whitespace-only partial (e.g. agent had only emitted some
-    // leading newlines) is also a no-op — no useful text to save.
-    #[test]
-    fn capture_partial_on_abort_noop_on_whitespace_only() {
-        let mut session = crate::session::Session::new("openrouter", "test-model", 100_000);
-        let baseline = session.messages.len();
-        let mut buf = String::from("   \n\n\t  ");
-        let stashed = capture_partial_on_abort(&mut buf, &mut session, "Esc", 0, &mut Vec::new());
-        assert!(!stashed);
-        assert_eq!(session.messages.len(), baseline);
-    }
-
-    // When tool calls ran in the same turn as the abort, the trailer
-    // must say so. The agent's preserved text only covers what was
-    // streamed via `AgentEvent::Token`; tool calls + results emitted
-    // separately are NOT in `response_buf`. Without this hint the
-    // next turn's LLM would see the partial as a definitive "this
-    // was the assistant's response" and could re-run side-effecting
-    // tool calls.
-    #[test]
-    fn capture_partial_on_abort_trailer_notes_tool_calls() {
-        let mut session = crate::session::Session::new("openrouter", "test-model", 100_000);
-        let mut buf = String::from("I deleted the file");
-        let stashed =
-            capture_partial_on_abort(&mut buf, &mut session, "Ctrl+C", 2, &mut Vec::new());
-        assert!(stashed);
-        let content = &session.messages.last().unwrap().content;
-        assert!(
-            content.contains("I deleted the file"),
-            "partial text dropped: {content:?}",
-        );
-        assert!(
-            content.contains("[interrupted by user (Ctrl+C);"),
-            "trailer prefix changed: {content:?}",
-        );
-        assert!(
-            content.contains("2 tool call"),
-            "trailer must mention tool call count: {content:?}",
-        );
-        assert!(
-            content.contains("not preserved"),
-            "trailer must warn that tool calls were not preserved: {content:?}",
-        );
-    }
-
-    // Single tool call uses singular phrasing — "1 tool call ran" not
-    // "1 tool calls ran". Tiny but the LLM is reading this verbatim.
-    #[test]
-    fn capture_partial_on_abort_trailer_handles_singular_tool_call() {
-        let mut session = crate::session::Session::new("openrouter", "test-model", 100_000);
-        let mut buf = String::from("Running tests now");
-        capture_partial_on_abort(&mut buf, &mut session, "Esc", 1, &mut Vec::new());
-        let content = &session.messages.last().unwrap().content;
-        assert!(
-            content.contains("1 tool call ran"),
-            "expected singular phrasing for 1 tool call: {content:?}",
-        );
-        assert!(
-            !content.contains("1 tool calls ran"),
-            "leaked plural for singular case: {content:?}",
-        );
-    }
-
-    // Rewind must sync tree.entries + message_store + leaf_id with
-    // the truncated `messages` slice. Without this, the tree
-    // references orphaned ids that no longer have content, and the
-    // leaf_id can point past the truncation. Subsequent fork /
-    // clone / save-load operations either fail or carry stale ids.
-    #[test]
-    fn rewind_truncates_tree_and_store_in_sync_with_messages() {
-        let mut session = crate::session::Session::new("p", "m", 100_000);
-        session.add_message(crate::session::MessageRole::User, "u1");
-        session.add_message(crate::session::MessageRole::Assistant, "a1");
-        session.add_message(crate::session::MessageRole::User, "u2");
-        session.add_message(crate::session::MessageRole::Assistant, "a2");
-        let baseline_tree = session.tree.entries.len();
-        assert_eq!(baseline_tree, 4, "fixture: 4 entries");
-
-        // Rewind back to the first user message (idx=1 in the
-        // reverse-order user list means the *first* user).
-        let mut renderer = crate::ui::renderer::Renderer::new().unwrap();
-        // idx=0 = "rewind through the most recent user prompt" → cut
-        // at the position of u2 → messages become [u1, a1].
-        let _ = rewind_session(&mut session, 0, &mut renderer);
-
-        // After rewind, messages has [u1, a1]; tree must agree.
-        assert_eq!(session.messages.len(), 2);
-        assert_eq!(
-            session.tree.entries.len(),
-            session.messages.len(),
-            "tree entries must match messages count; got tree={}, msgs={}",
-            session.tree.entries.len(),
-            session.messages.len(),
-        );
-        assert_eq!(
-            session.message_store.len(),
-            session.messages.len(),
-            "store must match messages count",
-        );
-        // Leaf points to the last remaining message.
-        let last_id = session.messages.last().unwrap().id.clone();
-        assert_eq!(
-            session.tree.leaf_id,
-            Some(last_id.clone()),
-            "leaf_id must anchor to the new tail",
-        );
-        // Every remaining message id has a tree entry + store entry.
-        for m in &session.messages {
-            assert!(
-                session.tree.entries.contains_key(&m.id),
-                "missing tree entry for {}",
-                m.id,
-            );
-            assert!(
-                session.message_store.contains_key(&m.id),
-                "missing store entry for {}",
-                m.id,
-            );
-        }
-    }
-
-    // The token accumulator on the abort path keeps `total_tokens`
-    // in sync with `total_estimated_tokens`. Both fields are
-    // TODO(cost-tracking) placeholders today but the inconsistency
-    // between Done/Interjected (which both update total_tokens) and
-    // abort (which didn't) made the abort case look like the agent
-    // produced zero tokens that turn.
-    #[test]
-    fn capture_partial_on_abort_keeps_total_tokens_in_sync() {
-        let mut session = crate::session::Session::new("openrouter", "test-model", 100_000);
-        let baseline_total = session.total_tokens;
-        let baseline_est = session.total_estimated_tokens;
-        let mut buf = String::from(
-            "A reasonably long partial response that should produce a non-zero token estimate.",
-        );
-        capture_partial_on_abort(&mut buf, &mut session, "Ctrl+C", 0, &mut Vec::new());
-        // Both fields advanced by the same amount (the stashed
-        // message's estimated_tokens). Without the parity fix, only
-        // total_estimated_tokens moved.
-        assert!(
-            session.total_estimated_tokens > baseline_est,
-            "total_estimated_tokens should advance on stash",
-        );
-        assert_eq!(
-            session.total_tokens.saturating_sub(baseline_total),
-            session.total_estimated_tokens.saturating_sub(baseline_est),
-            "total_tokens must advance in lockstep with total_estimated_tokens",
-        );
-    }
-
-    // Regression H1: lifecycle line for a failed task previously embedded the
-    // raw error string. Renderer.write_line splits on '\n', so a multi-line
-    // error broke the line layout (color reset, closing ']' on its own row).
-    // sanitize_single_line must collapse newlines into spaces.
-    #[test]
-    fn sanitize_replaces_newlines_with_space() {
-        let s = sanitize_single_line("line one\nline two\nline three", 100);
-        assert_eq!(s, "line one line two line three");
-        assert!(!s.contains('\n'));
-    }
-
-    #[test]
-    fn sanitize_replaces_carriage_return_and_tab() {
-        let s = sanitize_single_line("a\rb\tc", 100);
-        assert_eq!(s, "a b c");
-    }
-
-    // Regression: ANSI escape sequences (ESC = 0x1B) would otherwise be
-    // emitted verbatim and corrupt terminal state.
-    #[test]
-    fn sanitize_strips_ansi_escape() {
-        let s = sanitize_single_line("hello \x1b[31mred\x1b[0m world", 100);
-        assert!(!s.contains('\x1b'));
-        assert!(s.contains("hello"));
-        assert!(s.contains("world"));
-    }
-
-    // Other ASCII control chars (bell, backspace, etc.) are also stripped.
-    #[test]
-    fn sanitize_strips_other_controls() {
-        let s = sanitize_single_line("a\x07b\x08c\x00d", 100);
-        // Each control disappears; visible chars remain in order.
-        assert_eq!(s, "abcd");
-    }
-
-    #[test]
-    fn sanitize_truncates_at_char_limit() {
-        let s = sanitize_single_line(&"x".repeat(200), 50);
-        // 50 x's + ellipsis.
-        assert_eq!(s.chars().count(), 51);
-        assert!(s.ends_with('…'));
-    }
-
-    #[test]
-    fn sanitize_does_not_truncate_when_within_limit() {
-        let s = sanitize_single_line("hello", 100);
-        assert_eq!(s, "hello");
-        assert!(!s.ends_with('…'));
-    }
-
-    // Multibyte content counts by chars, not bytes, and remains intact.
-    #[test]
-    fn sanitize_handles_utf8_correctly() {
-        let s = sanitize_single_line("🦀🦀🦀\n🦀🦀", 100);
-        assert_eq!(s, "🦀🦀🦀 🦀🦀");
-    }
-
-    // Truncation at a multibyte boundary must produce valid UTF-8.
-    #[test]
-    fn sanitize_truncation_does_not_split_multibyte() {
-        let s = sanitize_single_line("🦀🦀🦀🦀🦀", 3);
-        // 3 emojis + ellipsis. No broken bytes.
-        assert_eq!(s.chars().count(), 4);
-        assert!(s.ends_with('…'));
-        // Round-trip as &str succeeds.
-        let _ = s.as_str();
-    }
-
-    #[test]
-    fn with_queue_hides_zero_count() {
-        // No interjections waiting → status line unchanged so the user
-        // doesn't see ambient "q:0" noise during normal operation.
-        let s = with_queue("ready".to_string(), 0);
-        assert_eq!(s, "ready");
-    }
-
-    #[test]
-    fn with_queue_appends_count() {
-        let s = with_queue("running".to_string(), 3);
-        assert!(s.ends_with("q:3"));
-        assert!(s.starts_with("running"));
-    }
-
-    /// User bug: `read` output containing a tab caused the chamber's
-    /// right border to drift right. `\t` has Unicode width 0 but the
-    /// terminal renders it as 4+ cells, so width-based padding
-    /// undercounted. The fix expands tabs to spaces (stop=4) before
-    /// measurement so the right `│` lands at the expected column.
-    #[test]
-    fn chamber_row_right_border_aligns_with_tabs() {
-        use unicode_width::UnicodeWidthStr;
-        let inner = 60;
-        // Three rows: no tab, one tab at start, tab embedded mid-line.
-        // After tab-expansion all should produce equal display width.
-        let rows = [
-            chamber_row("plain text", inner),
-            chamber_row("\tindented", inner),
-            chamber_row("2:\t(cd ..; make library)", inner),
-        ];
-        let widths: Vec<usize> = rows
-            .iter()
-            .map(|r| UnicodeWidthStr::width(r.as_str()))
-            .collect();
-        // All rows occupy exactly `inner + 4` cells (`│ ` + inner + ` │`).
-        let expected = inner + 4;
-        for (r, w) in rows.iter().zip(widths.iter()) {
-            assert_eq!(
-                *w, expected,
-                "chamber row width mismatch — content {r:?} measured {w} cells, want {expected}"
-            );
-        }
-        // Sanity: every row ends with `│` (right border didn't get
-        // pushed off into oblivion by under-padded tab).
-        for r in &rows {
-            assert!(r.ends_with('│'), "row {r:?} missing right border");
-        }
-    }
-
-    /// `chamber_row_with_bg` gets the same tab-expansion treatment so
-    /// diff `+`/`-` lines whose source uses tab indentation also
-    /// align correctly.
-    #[test]
-    fn chamber_row_with_bg_right_border_aligns_with_tabs() {
-        use unicode_width::UnicodeWidthStr;
-        let inner = 60;
-        let row = chamber_row_with_bg("+\tadded line", inner, 22);
-        // chamber_row_with_bg wraps content in SGR escapes; the
-        // visible width should still be inner + 4.
-        let visible = crate::ui::wrap::visible_width(&row);
-        assert_eq!(visible, inner + 4);
-        // Plain UnicodeWidthStr counts SGR payload too, but the
-        // visible-width helper from `wrap.rs` is the right tool.
-        // Sanity-only width assertion via the visible helper.
-        let _ = UnicodeWidthStr::width(row.as_str());
-        assert!(row.ends_with('│'));
-    }
-
-    /// Chat window switching: next / prev index math wraps correctly.
-    #[test]
-    fn chat_index_next_prev_wraps() {
-        // Simulate 3 chats (0=main, 1, 2), active=0.
-        let count = 3;
-        // Ctrl+N: next
-        assert_eq!((0 + 1) % count, 1);
-        assert_eq!((1 + 1) % count, 2);
-        assert_eq!((2 + 1) % count, 0); // wrap
-        // Ctrl+P: prev
-        assert_eq!((0 + count - 1) % count, 2); // wrap
-        assert_eq!((2 + count - 1) % count, 1);
-        assert_eq!((1 + count - 1) % count, 0);
-    }
-
-    /// Chat window switching: single chat is a no-op.
-    #[test]
-    fn chat_index_next_prev_one_chat_is_noop() {
-        let count = 1;
-        assert_eq!((0 + 1) % count, 0);
-        assert_eq!((0 + count - 1) % count, 0);
-    }
-}
+#[path = "mod_tests.rs"]
+mod tests;
